pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "graphql-client-generator"
version = "0.1.0"
description = "Generate typed Python clients from GraphQL schemas"
requires-python = ">=3.10"
dependencies = ["graphql-core>=3.2"]

[project.scripts]
graphql-client-generator = "graphql_client_generator.cli:main"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
    "requests",
    "mypy",
    "ruff",
]

[tool.hatch.envs.default.scripts]
test = "pytest tests/ -v --cov=graphql_client_generator --cov-report=term-missing"
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy graphql_client_generator"
check = ["lint", "typecheck", "test"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["graphql", "graphql.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
